rebaby_site/requirements.txt

Flask>=2.0
Flask-Login
Flask-WTF
Flask-SQLAlchemy
WTForms
Pillow
